Switch on travis's caching features.
We can re-use the following across builds:
* compiled OpenSSL libs
* compiled crate dependencies

Use specific folder caching because we've got a custom installation
of rust/cargo etc. so travis's `cargo` caching can't really be
expected to get that right. Move the install location of
OpenSSL to ensure that if the version is bumped, we won't get a bad
cache hit and silently fail to apply the upgrade.

If the rust toolchain gets bumped, it would be worth checking what
gains can be got by incremental compilation of the local crate.

+++ ci/run-docker.sh
@@ -145,20 +145,37 @@ case $TARGET in
     ;;
 esac
 
-mkdir -p target/$TARGET/openssl
-install=`pwd`/target/$TARGET/openssl/openssl-install
-out=`pwd`/target/$TARGET/openssl/openssl-$OPENSSL_VERS.tar.gz
-curl -o $out https://www.openssl.org/source/openssl-$OPENSSL_VERS.tar.gz
-sha256sum $out > $out.sha256
-test $OPENSSL_SHA256 = `cut -d ' ' -f 1 $out.sha256`
+install=`pwd`/target/$TARGET/openssl/openssl-install/$OPENSSL_VERS
 
-tar xf $out -C target/$TARGET/openssl
-(cd target/$TARGET/openssl/openssl-$OPENSSL_VERS && \
- CC=$OPENSSL_CC \
- AR=$OPENSSL_AR \
- $SETARCH ./Configure --prefix=$install no-dso $OPENSSL_OS $OPENSSL_CFLAGS -fPIC && \
- make -j4 && \
- make install)
+if [ -e $install ]; then
+  echo 'Using cached OpenSSL static libs'
+else
+  # Clean up any builds of previous versions from the cache
+  rm -rf $(dirname $install)/*
+
+  # If the build fails half way through it will be difficult to distinguish when the next run sees
+  # the cached version, so finalize the build atomically. We're linking statically so don't need to
+  # worry about using a different prefix at install time.
+  final_install_path=$install
+  install=$install-partial
+
+  mkdir -p target/$TARGET/openssl
+  out=`pwd`/target/$TARGET/openssl/openssl-$OPENSSL_VERS.tar.gz
+  curl -o $out https://www.openssl.org/source/openssl-$OPENSSL_VERS.tar.gz
+  sha256sum $out > $out.sha256
+  test $OPENSSL_SHA256 = `cut -d ' ' -f 1 $out.sha256`
+
+  tar xf $out -C target/$TARGET/openssl
+  (cd target/$TARGET/openssl/openssl-$OPENSSL_VERS && \
+   CC=$OPENSSL_CC \
+   AR=$OPENSSL_AR \
+   $SETARCH ./Configure --prefix=$install no-dso $OPENSSL_OS $OPENSSL_CFLAGS -fPIC && \
+   make -j4 && \
+   make install)
+
+   mv $install $final_install_path
+   install=$final_install_path
+fi
 
 # Variables to the openssl-sys crate to link statically against the OpenSSL we
 # just compiled above
